pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = "3.11.*"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "tf-keras (>=2.19.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

